+++ main/src/main/java/org/hibernate/tool/api/reveng/MetaDataDialectFactory.java
@@ -3,12 +3,6 @@ package org.hibernate.tool.api.reveng;
 import java.util.Properties;
 
 import org.hibernate.cfg.JDBCBinderException;
-import org.hibernate.cfg.reveng.dialect.H2MetaDataDialect;
-import org.hibernate.cfg.reveng.dialect.HSQLMetaDataDialect;
-import org.hibernate.cfg.reveng.dialect.JDBCMetaDataDialect;
-import org.hibernate.cfg.reveng.dialect.MySQLMetaDataDialect;
-import org.hibernate.cfg.reveng.dialect.OracleMetaDataDialect;
-import org.hibernate.cfg.reveng.dialect.SQLServerMetaDataDialect;
 import org.hibernate.dialect.Dialect;
 import org.hibernate.dialect.H2Dialect;
 import org.hibernate.dialect.HSQLDialect;
@@ -16,6 +10,12 @@ import org.hibernate.dialect.MySQLDialect;
 import org.hibernate.dialect.Oracle8iDialect;
 import org.hibernate.dialect.SQLServerDialect;
 import org.hibernate.internal.util.ReflectHelper;
+import org.hibernate.tool.internal.dialect.H2MetaDataDialect;
+import org.hibernate.tool.internal.dialect.HSQLMetaDataDialect;
+import org.hibernate.tool.internal.dialect.JDBCMetaDataDialect;
+import org.hibernate.tool.internal.dialect.MySQLMetaDataDialect;
+import org.hibernate.tool.internal.dialect.OracleMetaDataDialect;
+import org.hibernate.tool.internal.dialect.SQLServerMetaDataDialect;
 
 public class MetaDataDialectFactory {
