+++ main/src/main/java/org/hibernate/tool/internal/dialect/HSQLMetaDataDialect.java
@@ -1,4 +1,4 @@
-package org.hibernate.cfg.reveng.dialect;
+package org.hibernate.tool.internal.dialect;
 
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
