+++ main/src/main/java/org/hibernate/tool/internal/dialect/OracleMetaDataDialect.java
@@ -1,5 +1,5 @@
 
-package org.hibernate.cfg.reveng.dialect;
+package org.hibernate.tool.internal.dialect;
 
 import java.sql.Connection;
 import java.sql.PreparedStatement;
