+++ main/src/main/java/org/hibernate/tool/internal/dialect/ResultSetIterator.java
@@ -1,4 +1,4 @@
-package org.hibernate.cfg.reveng.dialect;
+package org.hibernate.tool.internal.dialect;
 
 import java.sql.ResultSet;
 import java.sql.SQLException;
