+++ main/src/main/java/org/hibernate/tool/internal/dialect/SQLServerMetaDataDialect.java
@@ -1,4 +1,4 @@
-package org.hibernate.cfg.reveng.dialect;
+package org.hibernate.tool.internal.dialect;
 
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
@@ -7,12 +7,13 @@ import java.util.HashMap;
 import java.util.Iterator;
 import java.util.Map;
 
-public class MySQLMetaDataDialect extends JDBCMetaDataDialect {
+/**
+ * 
+ * @author ddukker
+ *
+ */
+public class SQLServerMetaDataDialect extends JDBCMetaDataDialect {
 
-	/**
-	 * Based on info from http://dev.mysql.com/doc/refman/5.0/en/show-table-status.html
-	 * Should work on pre-mysql 5 too since it uses the "old" SHOW TABLE command instead of SELECT from infotable.
-	 */
 	public Iterator<Map<String, Object>> getSuggestedPrimaryKeyStrategyName(String catalog, String schema, String table) {
 		String sql = null;
 			try {			
@@ -22,7 +23,12 @@ public class MySQLMetaDataDialect extends JDBCMetaDataDialect {
 				
 				log.debug("geSuggestedPrimaryKeyStrategyName(" + catalog + "." + schema + "." + table + ")");
 				
-				sql = "show table status " + (catalog==null?"":" from " + catalog + " ") + (table==null?"":" like '" + table + "' ");
+				sql = "SELECT a.TABLE_CATALOG, a.TABLE_SCHEMA, a.TABLE_NAME as table_name, c.DATA_TYPE as data_type, b.CONSTRAINT_TYPE,  OBJECTPROPERTY(OBJECT_ID(a.TABLE_NAME),'TableHasIdentity') as hasIdentity " +
+						"FROM INFORMATION_SCHEMA.KEY_COLUMN_USAGE a " +
+						"INNER JOIN INFORMATION_SCHEMA.TABLE_CONSTRAINTS b on a.CONSTRAINT_NAME = b.CONSTRAINT_NAME " +
+						"INNER JOIN INFORMATION_SCHEMA.Columns c on a.TABLE_CATALOG = c.TABLE_CATALOG AND a.TABLE_SCHEMA = c.TABLE_SCHEMA AND a.TABLE_NAME = c.TABLE_NAME AND a.COLUMN_NAME = c.COLUMN_NAME " +
+						"WHERE a.TABLE_NAME='"+table+"' AND a.TABLE_SCHEMA='"+schema+"' AND a.TABLE_CATALOG='"+catalog+"' AND b.CONSTRAINT_TYPE = 'Primary key'";
+				
 				PreparedStatement statement = getConnection().prepareStatement( sql );
 				
 				final String sc = schema;
@@ -32,15 +38,23 @@ public class MySQLMetaDataDialect extends JDBCMetaDataDialect {
 					Map<String, Object> element = new HashMap<String, Object>();
 					protected Map<String, Object> convertRow(ResultSet tableRs) throws SQLException {
 						element.clear();
-						element.put("TABLE_NAME", tableRs.getString("NAME"));
+						element.put("TABLE_NAME", tableRs.getString("table_name"));
 						element.put("TABLE_SCHEM", sc);
 						element.put("TABLE_CAT", cat);
 						
-						String string = tableRs.getString("AUTO_INCREMENT");
-						if(string==null) {
+						String string = tableRs.getString("data_type");
+						
+						boolean bool = tableRs.getBoolean("hasIdentity");
+						if(string!=null) {
+							if(string.equalsIgnoreCase("uniqueidentifier")){
+								element.put("HIBERNATE_STRATEGY", "guid");
+							}else if(bool){
+								element.put("HIBERNATE_STRATEGY", "identity");
+							}else{
+								element.put("HIBERNATE_STRATEGY", null);
+							} 
+						}else {
 							element.put("HIBERNATE_STRATEGY", null);
-						} else {
-							element.put("HIBERNATE_STRATEGY", "identity");
 						}
 						return element;					
 					}
@@ -55,28 +69,6 @@ public class MySQLMetaDataDialect extends JDBCMetaDataDialect {
 				throw getSQLExceptionConverter().convert(e, "Could not get list of suggested identity strategies from database. Probably a JDBC driver problem. ", sql);		         
 			} 		
 		}
-	
-	@Override
-	public Iterator<Map<String,Object>> getTables(
-			String xcatalog, 
-			String xschema, 
-			String xtable) {
-	     // MySql JDBC Driver doesn't like 'null' values for the table search pattern, use '%' instead
-		return super.getTables(xcatalog, xschema, xtable != null ? xtable : "%");
-	}
 
-	public Iterator<Map<String, Object>> getColumns(
-			String xcatalog, 
-			String xschema, 
-			String xtable, 
-			String xcolumn) {
-	     // MySql JDBC Driver doesn't like 'null' values for the table and column search patterns, use '%' instead
-		return super.getColumns(
-				xcatalog, 
-				xschema, 
-				xtable != null ? xtable : "%",
-				xcolumn != null ? xcolumn : "%");
 	}
-	
-}
